LayoutTests/fast/css/font-systemFontID-parsing.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test.js"></script>
<div id="testDivInvalid" style="font: caption icon"></div>
<div id="testDivValid" style="font: caption"></div>
<script>
description("Tests 'font' CSS property parsing with system font values.")
shouldBeEqualToString("testDivInvalid.style.getPropertyValue('font')", "")
shouldBeEqualToString("testDivValid.style.getPropertyValue('font')", "caption")
shouldBeEqualToString("testDivInvalid.style.getPropertyValue('font-family')", "")
shouldBeEqualToString("testDivValid.style.getPropertyValue('font-family')", "")
</script>
</body>
</html>
