LayoutTests/fast/css/font-systemFontID-parsing-expected.txt

Tests 'font' CSS property parsing with system font values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDivInvalid.style.getPropertyValue('font') is ""
PASS testDivValid.style.getPropertyValue('font') is "caption"
PASS testDivInvalid.style.getPropertyValue('font-family') is ""
PASS testDivValid.style.getPropertyValue('font-family') is ""
PASS successfullyParsed is true

TEST COMPLETE
